State 'activity' created to manage every field on the form Element.
HTML generated to complete the UI, every input or element inside the Form Element has each 'value' and 'onChange' event.

'validateInputText' implemented with 'useMemo' to validate which category is selected to change between two values in the Input Submit.

'handleChangeSelect' func to set values to the State and parsing 'category' and 'calories' to a number.

'isValidActivity' validates if 'name' and 'calories' fields are empty.

'handleSubmit' waiting to implement his functionality.